Añadido modificar acciones portfolio desde admin y aumento decimales a 12 en tasas de cambio

+++ server/models/Operation.js
@@ -41,7 +41,7 @@ const Operation = sequelize.define('Operation', {
     defaultValue: 'EUR'
   },
   exchangeRate: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(18, 12), // 18 dígitos totales, 12 decimales
     allowNull: false,
     defaultValue: 1
   },
